+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - You can now use shorthands and `web/native: {}` scopes inside `zacs:style={}` attribute. Note that this only works if an object literal is passed. You can transform any such style object with `zacs._experimental_resolve({ ... })`
 - `<zacs.view />` and `<zacs.text />` anonymous ZACS elements
 - You can now use `...mixin` syntax in stylesheets instead of `_mixin: mixin`
+- On React Native, stylesheet mixins can now be arbitrary expressions and ZACS will generate more efficient representation
 
 **Performance**